[ARM] msm: kgsl: Remove support for mmap from the kgsl driver.
The mmap(2) functionality of kgsl is no longer used, and it had a
bug that would crash the kernel when mmap(2) was called with
odd parameters.  So we have removed the mmap functionality.
If for some reason in the future you want to resurrect this,
make sure memdesc is properly initialized in all paths through
the kgsl_mmap() routine.